requirements.txt
requests==2.32.3
streamlit==1.38.0
folium==0.17.0
pandas==2.2.2
pyrebase4==4.8.0